TAL, TPB, TCM, TCO, TCR, TDA TDY, TEN, TFT, TIM, TKE, TLA, TLE, TMT, TOA, TOF, TOL, TOR, TOT, TP1, TP2, TP3, TP4, TPA, TPB, TRC, TRD, TRK, TSI, TSS, TT1, TT2, TT3, TYE Tag Frames.

+++ src/main/java/com/aurawin/mp3/frame/Header.java
@@ -11,7 +11,7 @@ public abstract class Header {
     public byte VersionMinor;
     public  Reader Reader;
     public long StreamStart;
-    public  short Length;
+    public  long Length;
 
     public void Reset(){
         VersionMajor =0;
