+++ src/main/java/com/aurawin/mp3/frame/tag/process/V20/CNT.java
@@ -11,7 +11,7 @@ public class CNT extends Processor implements Process {
     public boolean process( MemoryStream Stream){
         Owner.Data=Owner.pldCNT;
         Owner.Reader.TagFrame.Kind=fPlayCounter;
-        Owner.pldCNT.Counter= Stream.readWhole(Owner.Reader.TagFrame.Length);
+        Owner.pldCNT.Counter= Stream.readWhole((int) Owner.Reader.TagFrame.Length);
 
         Owner.Length=Owner.Reader.TagFrame.Header.Length;
         return true;
